Emscripten: very briefly working
This package uses alot of threads, which Emscripten does not to support very well with dynamic linking that Lua does.

The optimal solution would be dynamic linking and threads. But the current state forces us to either go DL or threads. So either find 100% Lua libs, or another way to prevent using dlopen(). The other solution is removing threading usage (which is this commit attempts, but fails very badly).